Move file rendering from UMLRoot to Diagram
